+++ apps/gotask_ui/src/app/(portal)/goals/projectid/[projectId]/components/projectGoal/projectGoals.tsx
@@ -3,7 +3,7 @@ import { Grid, Box, Typography } from "@mui/material";
 import { useTranslations } from "next-intl";
 import { LOCALIZATION } from "@/app/common/constants/localization";
 import { ProjectGoalsProps } from "@/app/(portal)/goals/projectid/[projectId]/interface/projectGoal";
-import GoalCard from "@/app/(portal)/goals/projectid/[projectId]/components/projectGoal/projectGoalCards";
+import GoalCard from "./goalItem";
 
 const ProjectGoals: React.FC<ProjectGoalsProps> = ({ projectGoals, projectId, handleEditGoal }) => {
   const filteredGoals = projectGoals.filter((goal) => goal.projectId === projectId);
